+++ upgrade.php
@@ -63,6 +63,7 @@ function agenda_upgrade()
         case '0.3.1' :
         case '0.3.2' :
         case '0.4.0' :
+        case '0.5.0' :
 
         default:
             agenda_update_config();