+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 Agenda Plugin ChangeLog
 
+## v0.6.0 (November 23, 2017)
+ - You can now edit all attributes of an event series
+ - Use datetimepicker plugin for pop-up calendar
+ - Do not allow end date to be prior to begin date
+ - Code consolidations and cleanup
+
 ## v0.5.0 (November 16, 2017)
  - Add print option
  - New configuration setting to enable print button / feature